fix: compiler warnings

+++ Singular/subexpr.cc
@@ -1065,16 +1065,16 @@ void * sleftv::Data()
         idhdl h=(idhdl)data;
         return  ((idhdl)h->data.ustring)->data.ustring;
       }
-      case VECHO:      return (void *)si_echo;
-      case VPRINTLEVEL:return (void *)printlevel;
-      case VCOLMAX:    return (void *)colmax;
-      case VTIMER:     return (void *)getTimer();
-      case VRTIMER:    return (void *)getRTimer();
-      case VOICE:      return (void *)(myynest+1);
-      case VMAXDEG:    return (void *)Kstd1_deg;
-      case VMAXMULT:   return (void *)Kstd1_mu;
-      case TRACE:      return (void *)traceit;
-      case VSHORTOUT:  return (void *)(currRing != NULL ? currRing->ShortOut : 0);
+      case VECHO:      return (void *)(long)si_echo;
+      case VPRINTLEVEL:return (void *)(long)printlevel;
+      case VCOLMAX:    return (void *)(long)colmax;
+      case VTIMER:     return (void *)(long)getTimer();
+      case VRTIMER:    return (void *)(long)getRTimer();
+      case VOICE:      return (void *)(long)(myynest+1);
+      case VMAXDEG:    return (void *)(long)Kstd1_deg;
+      case VMAXMULT:   return (void *)(long)Kstd1_mu;
+      case TRACE:      return (void *)(long)traceit;
+      case VSHORTOUT:  return (void *)(long)(currRing != NULL ? currRing->ShortOut : 0);
       case VMINPOLY:
         if ( (currRing != NULL)  && nCoeff_is_algExt(currRing->cf) && !nCoeff_is_GF(currRing->cf))
         {
